# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Conjur API
  description: OpenAPI specification for Conjur - compatible with CyberArk Conjur
  version: "5.0.0"
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Basic authentication with username and API key
    bearerAuth:
      type: http
      scheme: bearer
      description: Conjur access token
  schemas:
...
    AccountCreateResponse:
      type: object
      properties:
        id:
          type: string
        api_key:
          type: string
...
  parameters:
    Account:
      name: account
      in: path
      required: true
      schema:
        type: string
      description: Organization account name
    AccountId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Account ID
...
    Login:
      name: login
      in: path
      required: true
      schema:
        type: string
      description: URL-encoded login name
    ServiceId:
      name: service_id
      in: path
      required: true
      schema:
        type: string
      description: Authenticator service ID
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      summary: List all accounts
      tags: [accounts]
      security: []
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

    post:
      operationId: createAccount
      summary: Create a new account
      tags: [accounts]
      security: []
      parameters:
        - name: id
          in: query
          schema:
            type: string
          description: Account name
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountCreateResponse"
        "400":
          description: Invalid account name
        "409":
          description: Account already exists
  /accounts/{id}:
    delete:
      operationId: deleteAccount
      summary: Delete an account
      tags: [accounts]
      security: []
      parameters:
        - $ref: "#/components/parameters/AccountId"
      responses:
        "204":
          description: Account deleted
        "404":
          description: Account not found

  # ========== AUTHENTICATION ==========
...
  /authn/{account}/api_key:
    put:
      operationId: rotateApiKey
      summary: Rotate API key
      tags: [authentication]
      security:
        - basicAuth: []
      parameters:
        - $ref: "#/components/parameters/Account"
        - name: role
          in: query
          schema:
            type: string
          description: Target role to rotate key for
      responses:
        "200":
          description: New API key
          content:
            text/plain:
              schema:
                type: string
        "401":
          description: Invalid credentials
        "403":
          description: Forbidden
  /authn/{account}/password:
    put:
      operationId: changePassword
      summary: Change password
      tags: [authentication]
      security:
        - basicAuth: []
      parameters:
        - $ref: "#/components/parameters/Account"
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
              description: New password
      responses:
        "204":
          description: Password changed
        "401":
          description: Invalid credentials
        "403":
          description: Hosts cannot change passwords
  /authn/{account}/{login}/authenticate:
    post:
      operationId: authenticate
      summary: Authenticate and get access token
      tags: [authentication]
      security: []
      parameters:
        - $ref: "#/components/parameters/Account"
        - $ref: "#/components/parameters/Login"
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
              description: API key
      responses:
        "200":
          description: Access token
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Invalid credentials
  /authn-jwt/{service_id}/{account}/authenticate:
    post:
      operationId: authenticateJwt
      summary: Authenticate using JWT
      tags: [authentication]
      security: []
      parameters:
        - $ref: "#/components/parameters/ServiceId"
        - $ref: "#/components/parameters/Account"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                jwt:
                  type: string
      responses:
        "200":
          description: Access token
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Invalid JWT

  # ========== SECRETS ==========
...